+++ twoaxistracking/__init__.py
@@ -1,7 +1,7 @@
 # Import of functions that should be accessible from the package top-level
 from .layout import generate_field_layout  # noqa: F401
 from .shading import shaded_fraction  # noqa: F401
-from .twoaxistrackerfield import TwoAxisTrackerField  # noqa: F401
+from .trackerfield import TrackerField  # noqa: F401
 
 try:
     from shapely.geos import lgeos  # noqa: F401