+++ twoaxistracking/plotting.py
@@ -4,7 +4,6 @@ from matplotlib import patches
 from shapely import geometry
 import matplotlib.colors as mcolors
 from matplotlib import cm
-import numpy as np
 
 
 def _plot_field_layout(X, Y, Z, min_tracker_spacing):
@@ -48,7 +47,7 @@ def _polygons_to_patch_collection(geometries, **kwargs):
         geometries = [geometries]
     elif isinstance(geometries, geometry.MultiPolygon):
         geometries = list(geometries.geoms)
-    exteriors = [patches.Polygon(np.array(g.exterior)) for g in geometries]
+    exteriors = [patches.Polygon(g.exterior.coords) for g in geometries]
     path_collection = collections.PatchCollection(exteriors, **kwargs)
     return path_collection